Most functionality of the ALUs and Data Mem Complete
It compiles, but it has not been tested. One thing we will have to figure is how the jz8 and jnz8 instructions will work.
Also, I realized every instruction that is not lf, li, or st needs to pass through the alu. To accomplish this, I created a case
for every remaining instruction in the alu. Some things will have to be completed at writeback however.
Have not hooked up the ALU to the processor.
The wiring scheme I used is: 
     If instruction is one param: in1 = acc,        in2 = register or imm8.
     If instruction is two param, in1 = register, in2 = imm8.